Add edit tags batch 3
