+++ Wikipedia/Code/WikiTextSectionUploader.h
@@ -28,7 +28,6 @@ extern NSString *const NSErrorUserInfoDisplayError;
              baseRevID:(nullable NSNumber *)baseRevID
              captchaId:(nullable NSString *)captchaId
            captchaWord:(nullable NSString *)captchaWord
-        editSummaryTag:(nullable NSString *)editSummaryTag
               editTags:(nullable NSArray<NSString *> *)editTags
             completion:(void (^)(NSDictionary * _Nullable result, NSError * _Nullable error))completion;
